+++ app/app.ts
@@ -31,9 +31,11 @@ const App = Application.extend({
         handbook: {
             dependencies: {
                 services: [
-                    'router',
                     'analytics',
                     'i18n',
+                    'router',
+                    'store',
+                    'toast',
                 ],
             },
         },