+++ app/settings/tokens/edit/controller.ts
@@ -12,21 +12,34 @@ export default class SettingsTokensEditController extends Controller {
 
     deleteModalShown: boolean = false;
 
-    @readOnly('model.taskInstance.value.token')
+    @readOnly('model.taskInstance.value')
     token?: Token;
-
-    @readOnly('model.taskInstance.value.tokenId')
-    tokenId?: string;
 
     @action
     tokenSaved() {
-        // Analytics handled by x-token-form and x-delete-modal
+        // Analytics handled by x-token-form
         this.router.transitionTo('settings.tokens');
     }
 
-    clearTokenId() {
-        if (this.token) {
-            this.token.set('tokenId', undefined);
+    @action
+    async deleteToken() {
+        // Analytics handled by delete-button
+        await this.token!.destroyRecord();
+
+        this.router.transitionTo('settings.tokens');
+    }
+
+    @action
+    refresh() {
+        this.clearTokenValue();
+
+        // Send action to route
+        this.send('refreshRoute');
+    }
+
+    clearTokenValue() {
+        if (this.token && this.token.tokenValue) {
+            this.token.unloadRecord();
         }
     }
 }