+++ mirage/scenarios/default.ts
@@ -36,7 +36,8 @@ export default function(server: Server) {
         server.create('contributor', { node, users: currentUser, index: 11 });
     }
     server.createList('institution', 20);
-    server.createList('token', 7);
+    server.createList('token', 23);
+    server.createList('scope', 5);
     server.loadFixtures('registration-schemas');
     registerNodeMultiple(server, nodes[0], 12, {
         currentUserPermissions: Object.values(Permission),