+++ mirage/views/user.ts
@@ -15,6 +15,9 @@ export function userNodeList(schema: Schema, request: Request, handlerContext: H
     return json;
 }
 
-export function userList(schema: Schema, request: Request) {
-    return schema.users.where(user => filter(user, request));
+export function userFileList(schema: Schema, request: Request, handlerContext: HandlerContext) {
+    const user = schema.users.find(request.params.id);
+    const files = user.schema.files.all().models.map((file: any) => handlerContext.serialize(file).data);
+    const json = process(schema, request, handlerContext, files, { defaultSortKey: 'date_modified' });
+    return json;
 }