+++ types/ember-cli-mirage/index.d.ts
@@ -1,4 +1,10 @@
+import { Document } from 'osf-api';
+
 export { default as faker } from 'faker';
+
+declare global {
+    const server: Server; // TODO: only in tests?
+}
 
 type ID = number | string;
 
@@ -42,7 +48,7 @@ interface ModelInstanceShared<T> {
     toString(): string;
 }
 
-export type ModelInstance<T> = ModelInstanceShared<T> & Model<T>;
+export type ModelInstance<T = AnyAttrs> = ModelInstanceShared<T> & Model<T>;
 
 interface Collection<T> {
     models: Array<ModelInstance<T>>;
@@ -72,6 +78,7 @@ export interface Schema {
 
 export interface Request {
     requestBody: any;
+    url: string;
     params: {
         [key: string]: string | number,
     };
@@ -125,6 +132,9 @@ function handlerDefinition(
 /* tslint:enable unified-signatures */
 
 export interface Server {
+    schema: Schema;
+    db: Database;
+
     namespace: string;
     timing: number;
     logging: boolean;
@@ -137,8 +147,6 @@ export interface Server {
     put: typeof handlerDefinition;
     patch: typeof handlerDefinition;
     del: typeof handlerDefinition;
-
-    schema: Schema;
 
     resource(resourceName: string, options?: { only?: string[], except?: string[], path?: string }): void;
 
@@ -189,7 +197,9 @@ export function trait<O extends TraitOptions>(options: O): Trait<O>;
 export function association(...args: any[]): unknown;
 
 export type FactoryAttrs<T> = {
-    [P in keyof T]?: T[P] | ((...args: any[]) => T[P]);
+    [P in keyof T]?: T[P] | ((index: number) => T[P]);
+} & {
+    afterCreate?(newObj: any, server: Server): void;
 };
 
 export class FactoryClass {
@@ -198,14 +208,15 @@ export class FactoryClass {
 
 export const Factory: FactoryClass;
 
-declare global {
-    const server: Server; // TODO: only in tests?
-}
-
 export class JSONAPISerializer {
     request!: Request;
 
     keyForAttribute(attr: string): string;
-
+    keyForCollection(modelName: string): string;
+    keyForModel(modelName: string): string;
     keyForRelationship(relationship: string): string;
+    typeKeyForModel(model: ModelInstance): string;
+
+    serialize(object: ModelInstance, request: Request): SingleResourceDocument;
+    normalize(json: any): any;
 }