+++ types/osf-api.d.ts
@@ -33,8 +33,9 @@ export interface PaginatedMeta extends BaseMeta {
 }
 
 export interface RootMeta extends BaseMeta {
-    activeFlags: string[];
+    active_flags: string[];
     message: string;
+    version: string;
     current_user?: { data: UserResource };
 }
 
@@ -47,6 +48,7 @@ export interface Resource extends JSONAPI.ResourceObject {
     id: string | number;
     relationships?: Relationships;
     embeds?: Embeds;
+    links?: NormalLinks;
 }
 
 export interface UserResource extends Resource {
@@ -81,4 +83,15 @@ export interface RelatedLinkMeta {
     count?: number;
 }
 
+export interface NormalLinks extends JSONAPI.Links {
+    info?: Link | null;
+    self?: Link | null;
+    move?: Link | null;
+    upload?: Link | null;
+    download?: Link | null;
+    delete?: Link | null;
+    self?: Link | null;
+    html?: Link | null;
+    profile_image?: Link | null;
+}
 /* eslint-enable no-use-before-define,camelcase */